refactor: image_scales interfaces and dependencies - refs #254889
