+++ eea/volto/policy/image_scales/__init__.py
@@ -1,0 +1,1 @@
+""" Image scales fallback for Plone 5 """
